pyproject.toml
[project]
name = "deepagents"
version = "0.0.11"
description = "General purpose 'deep agent' with sub-agent spawning, todo list capabilities, and mock file system. Built on LangGraph."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11,<4.0"
dependencies = [
    "langgraph>=1.0.0a3",
    "langchain-anthropic>=0.1.23",
    "langchain>=1.0.0a12",
    "langgraph-prebuilt>=0.7.0a2",
    "langchain-core>=1.0.0a6",
    "python-dotenv>=1.1.1",
    "tavily-python>=0.7.12",
    "aiohttp>=3.13.0",
    "beautifulsoup4>=4.14.2",
    "firecrawl-py>=4.4.0",
    "requests>=2.32.5",
    "pydantic>=2.12.2",
    "lxml>=6.0.2",
    "trafilatura>=1.6.1",
    "langchain-openai>=0.3.34",
    "brotli>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "build",
    "twine",
]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["deepagents", "tests"]
[tool.setuptools.package-dir]
"deepagents" = "src/deepagents"
"tests" = "tests"

[tool.setuptools.package-data]
"*" = ["py.typed"]
